requirements.txt
chromadb==1.0.9
grandalf==0.8
langchain==0.3.25
langchain_community==0.3.24
langchain-openai==0.3.17
langgraph==0.4.5
langgraph-checkpoint-sqlite==2.0.7
pydantic==2.11.4
python-dotenv==1.1.0
streamlit==1.45.1
PyPDF2==3.0.1
pymongo[srv]==3.11